requirements.txt
discord.py
aiohttp


